Final app includes both traditional REST and GraphQL endpoints
